requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.24
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
constantly==15.1.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
diff-match-patch==20121119
Django==1.11
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-reversion==3.0.1
django-simple-captcha==0.5.9
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.17.1
httpie==0.9.9
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
jieba==0.39
Jinja2==2.10
kombu==3.0.37
lxml==4.2.3
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.4
parsel==1.5.1
pefile==2018.8.8
Pillow==5.0.0
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycryptodomex==3.7.0
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyMySQL==0.8.1
pyOpenSSL==18.0.0
python-alipay-sdk==1.8.0
python-ptrace==0.9.3
pytz==2018.4
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.8
tablib==0.12.1
Twisted==18.9.0
unicodecsv==0.14.1
uritemplate==3.0.0
urljoin==1.0.0
urllib3==1.23
urlopen==1.0.0
vine==1.1.4
w3lib==1.19.0
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.6.0
